+++ controller/pkg/extensions/validate.go
@@ -8,14 +8,13 @@ import (
 
 	"github.com/kong/gateway-operator/pkg/consts"
 
-	operatorv1beta1 "github.com/kong/kubernetes-configuration/api/gateway-operator/v1beta1"
 	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 )
 
 // ValidateExtensions validates the extensions referenced by the provided DataPlane and returns
 // a condition indicating if the extensions are supported or not.
-func ValidateExtensions(dataplane *operatorv1beta1.DataPlane) *metav1.Condition {
-	if len(dataplane.Spec.Extensions) == 0 {
+func validateExtensions[t ExtendableT](obj t) *metav1.Condition {
+	if len(obj.GetExtensions()) == 0 {
 		return nil
 	}
 
@@ -24,16 +23,16 @@ func ValidateExtensions(dataplane *operatorv1beta1.DataPlane) *metav1.Condition 
 		Type:               string(consts.AcceptedExtensionsType),
 		Reason:             string(consts.AcceptedExtensionsReason),
 		Message:            "All extensions are accepted",
-		ObservedGeneration: dataplane.Generation,
+		ObservedGeneration: obj.GetGeneration(),
 		LastTransitionTime: metav1.Now(),
 	}
 	var messageBuilder strings.Builder
-	for i, ext := range dataplane.Spec.Extensions {
+	for i, ext := range obj.GetExtensions() {
 		if ext.Group != konnectv1alpha1.SchemeGroupVersion.Group || ext.Kind != konnectv1alpha1.KonnectExtensionKind {
 			buildMessage(&messageBuilder, fmt.Sprintf("Extension %s/%s is not supported", ext.Group, ext.Kind))
 			continue
 		}
-		for j, ext2 := range dataplane.Spec.Extensions {
+		for j, ext2 := range obj.GetExtensions() {
 			if i != j {
 				if ext.Group == ext2.Group && ext.Kind == ext2.Kind {
 					message := fmt.Sprintf("Extension %s/%s is duplicated", ext.Group, ext.Kind)