+++ modules/manager/controller_setup.go
@@ -184,8 +184,11 @@ func setupIndexes(ctx context.Context, mgr manager.Manager, cfg Config) error {
 			}
 		}
 		if cfg.KonnectControllersEnabled {
-			if err := index.DataPlaneOnDataPlaneKonnecExtension(ctx, mgr.GetCache()); err != nil {
+			if err := index.ExtendableOnKonnectExtension(ctx, mgr.GetCache(), &operatorv1beta1.DataPlane{}); err != nil {
 				return fmt.Errorf("failed to setup index for DataPlanes on KonnectExtensions: %w", err)
+			}
+			if err := index.ExtendableOnKonnectExtension(ctx, mgr.GetCache(), &operatorv1beta1.ControlPlane{}); err != nil {
+				return fmt.Errorf("failed to setup index for ControlPlanes on KonnectExtensions: %w", err)
 			}
 		}
 	}
@@ -429,6 +432,7 @@ func SetupControllers(mgr manager.Manager, c *Config) (map[string]ControllerDef,
 				ClusterCASecretNamespace: c.ClusterCASecretNamespace,
 				ClusterCAKeyConfig:       clusterCAKeyConfig,
 				DevelopmentMode:          c.DevelopmentMode,
+				KonnectEnabled:           c.KonnectControllersEnabled,
 			},
 		},
 		// DataPlane controller